working on type 0 struct parsing
