+++ Makefile
@@ -2,7 +2,7 @@ CXX = g++
 CXXFLAGS = -std=c++20 -Wall -Wextra -Iinclude
 CXXFLAGS += -I./include
 
-SRC = $(wildcard src/*.cpp src/structures/*.cpp)
+SRC = $(wildcard src/*.cpp src/types/*.cpp)
 
 SMBIOS-Parser: $(SRC)
 	$(CXX) $(CXXFLAGS) $^ -o $@
