+++ src/types/TypeOther.cpp
